Implement the replication algorithm in the Node
We have implemented the replication algorithm in the node. As we
are using a Merkle tree to avoid inconsitency between the same
nodes, a method was added for entropy. As the problem on conciliating
the merkle trees is a big implementation we will implement it later
if it is needed.